error and exception handling

+++ src/main/java/io/github/beletatti/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.github.beletatti.libraryapi.service;
 
+import io.github.beletatti.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.beletatti.libraryapi.model.Autor;
 import io.github.beletatti.libraryapi.repository.AutorRepository;
+import io.github.beletatti.libraryapi.repository.LivroRepository;
+import io.github.beletatti.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,12 +15,17 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository) {
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository) {
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
 
     }
@@ -26,6 +34,7 @@ public class AutorService {
         if(autor.getId() == null){
             throw new IllegalArgumentException("Para atualizar, é necessário que o autor já esteja salvo na base.");
         }
+        validator.validar(autor);
         repository.save(autor);
 
     }
@@ -35,6 +44,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um Autor possui livros cadastrados.");
+        }
         repository.delete(autor);
     }
 
@@ -52,5 +64,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
